mixin rewrotten with kotlin
