--- src/main/java/com/dainxt/weaponthrow/mixins/IPlayerEntityMixin.java
@@ -1,9 +1,0 @@
-package com.dainxt.weaponthrow.mixins;
-
-import com.dainxt.weaponthrow.capabilities.PlayerThrowData;
-
-public interface IPlayerEntityMixin {
-	public void setThrowPower(PlayerThrowData value);
-	
-	public PlayerThrowData getThrowPower();
-}
