+++ src/main/java/com/dainxt/weaponthrow/mixins/IPlayerEntityMixin_java.cpp
@@ -1,0 +1,9 @@
+package com.dainxt.weaponthrow.mixins;
+
+import com.dainxt.weaponthrow.capabilities.PlayerThrowData;
+
+public interface IPlayerEntityMixin {
+	public void setThrowPower(PlayerThrowData value);
+	
+	public PlayerThrowData getThrowPower();
+}
