+++ src/main/java/com/dainxt/weaponthrow/mixins/IPlayerEntityMixin_kt.cpp
@@ -1,0 +1,7 @@
+import com.dainxt.weaponthrow.capabilities.PlayerThrowData
+
+interface IPlayerEntityMixin {
+    fun setThrowPower(value: PlayerThrowData)
+
+    fun getThrowPower(): PlayerThrowData
+}
